pyproject.toml
[project]
name = "agent2agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nest-asyncio>=1.6.0",
    "dotenv>=0.9.9",
    "rich>=14.3.4",
    "litellm>=1.83.4",
    "a2a-sdk>=0.3.26",
    "pypdf>=6.10.0",
    "uvicorn>=0.44.0",
]
